% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codonComparisons.R
\docType{methods}
\name{codonCompare}
\alias{codonCompare}
\alias{MCUFD}
\alias{MCUFD,codonFreq-method}
\title{Calculate measures of codon usage similarity.}
\usage{
MCUFD(cFobj, exclude = character(length = 0), minlen = 3)

\S4method{MCUFD}{codonFreq}(cFobj, exclude = character(length = 0),
  minlen = 3)
}
\arguments{
\item{cFobj}{An object of class \code{codonFreq}.}

\item{exclude}{A character vector of codons to be excluded from comparisons.}

\item{minlen}{Numeric, the minimum length of sequence (in nucleotides) to be
included in the analysis (default is 3).}
}
\value{
A data.frame containing values which indicate the degree of codon
   usage similarity between the sequences used to construct the
   \code{codonFreq} object and those used to construct the reference
   database. The reference database entries are in rows and the
   \code{codonFreq} sequences are in columns.
}
\description{
Compare codon usage between sequences in a \code{codonFreq} object with those
   in a reference database.
Mean codon usage frequency difference (\code{MCUFD}) is calculated as
   described in \href{https://goo.gl/7sGZeE}{Stedman et al. (2013)}.
}
\examples{
## To be added

}
