% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codondiffR.R
\docType{package}
\name{codondiffR}
\alias{codondiffR}
\title{codondiffR: inter-taxon comparison of codon usage}
\description{
An R package for the comparative analysis of codon usage in a set of
user-defined sequences with those in reference taxa. Allows the calculation
and visualisation of codon usage metrics and statistical analysis of
differences in these metrics between taxa.
}
