% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-codonFreq.R
\docType{methods}
\name{codonFreq}
\alias{codonFreq}
\title{Create new objects of class \code{codonFreq}.}
\usage{
\S4method{codonFreq}{DNAStringSet}(object)
}
\arguments{
\item{object}{An object of class \code{DNAStringSet}.}
}
\value{
a \code{codonFreq} object.
}
\description{
Create new objects of class \code{codonFreq}.
}
