% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{codonLDA-class}
\alias{codonLDA-class}
\title{S4 class for storing relative codon frequencies in user-supplied sequences.}
\description{
S4 class for storing relative codon frequencies in user-supplied sequences.
}
\section{Slots}{

\describe{
\item{\code{seqID}}{Character vector of sequence identifiers, derived from Fasta
identifier lines.}

\item{\code{codon_perc}}{Matrix containing the relative frequencies of codons; each
sequence is a row and each codon is a column.}

\item{\code{ncod}}{Numeric vector of sequence lengths (in codons).}
}}

