% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codonFreq-class.R
\docType{methods}
\name{getFreqs}
\alias{getFreqs}
\alias{getFreqs,codonFreq-method}
\title{Get relative frequencies from a \code{codonFreq} object.}
\usage{
getFreqs(object)

\S4method{getFreqs}{codonFreq}(object)
}
\arguments{
\item{object}{An object of class \code{codonFreq}.}
}
\value{
Matrix, frequencies of codons in the \code{codonFreq} object.
   Codons are in columns and input sequences are in rows.
}
\description{
Returns the relative frequencies of codons per sequence in a \code{codonFreq}
   object.
}
\section{Methods (by class)}{
\itemize{
\item \code{codonFreq}: 
}}

