% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/LDA.R
\docType{methods}
\name{LDA}
\alias{LDA}
\alias{bootstrap_LDA}
\title{Discriminant analysis of codon usage by taxon}
\usage{
LDA(cFobj, exclude = character(length = 0), minlen = 600,
  norm = FALSE, trans = FALSE, propTrain = 0.8, rank = "Phylum",
  corCut = 0.9)

bootstrap_LDA(rep = 100, trans = FALSE, propTrain = 0.8,
  rank = "Phylum", exclude = character(length = 0), minlen = 600,
  corCut = 0.9)

\S4method{LDA}{codonFreq}(cFobj, exclude = character(length = 0),
  minlen = 600, norm = FALSE, trans = FALSE, propTrain = 0.8,
  rank = "Phylum", corCut = 0.9)

\S4method{bootstrap_LDA}{numeric}(rep = 100, trans = FALSE,
  propTrain = 0.8, rank = "Phylum", exclude = character(length = 0),
  minlen = 600, corCut = 0.9)
}
\arguments{
\item{cFobj}{An object of class \code{codonFreq}.}

\item{exclude}{A character vector of codons to be excluded from comparisons.}

\item{minlen}{Numeric, the minimum length of sequence (in codons) to be
included in the analysis. Default = 500.}

\item{norm}{Logical, should the codon abundances be normalised? If TRUE, a
normalisation step will be performed, such that the proportions for each
codon sum to 1. Default = FALSE.}

\item{trans}{Logical; if true, a Box-Cox transformation will be applied to
the data. Default = FALSE.}

\item{propTrain}{Numeric, proportion of the reference database to use for
the LDA training set (must be in the range [0,1]). Default = 0.8.}

\item{rank}{Character, taxonomic rank to be used for categorisation. Options
are "Domain", "Kingdom", and "Phylum". Default = "Phylum".}

\item{corCut}{Numeric, correlation cutoff used for dropping codons that
exhibit collinearity. Default = 0.9.}

\item{rep}{Numeric, number of bootstrap replicates to perform (default =
100).}

\item{propTrain}{Numeric, proportion of the reference database to use for
the LDA training set (must be in the range [0,1]). Default = 0.8.}

\item{rank}{Character, taxonomic rank to be used for categorisation. Options
are "Domain", "Kingdom", and "Phylum". Default = "Phylum".}

\item{corCut}{Numeric, correlation cutoff used for dropping codons that
exhibit collinearity. Default = 0.9.}
}
\value{
x

A named list of length 3. The elements are "Codons" (the codons used
   for model construction, aftering the filtering criteria have been
   applied), "Taxa" (taxa used for model construction/testing), and
   "Accuracy" (the proportion of accurate classifications using the test
   subset of the data).
}
\description{
Perform linear discriminant analysis (LDA) on codon usage in a reference
   database and use it to classify sequences of unknown taxonomic affinity.
   Data can be optionally scaled using the
   \href{https://goo.gl/qzN63P}{Box-Cox power transformation}.

Perform random sampling using different subsets of the reference database
   to assess the impact on model accuracy.
}
\examples{
## To be added

   exclCod <- c("ATT", "TGT")
   boot <- bootstrap_LDA(
       rep = 1, propTrain = 0.9, trans = FALSE, rank = "Phylum",
       exclude = exclCod, minlen = 600, corCut = 0.9
   )
   names(boot)
   mean(boot$Accuracy)

}
