% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/MCUFD.R
\docType{methods}
\name{mcufd}
\alias{mcufd}
\alias{mcufd_enrich}
\alias{MCUFD}
\title{MCUFD enrichment testing}
\usage{
mcufd(cFobj, exclude = character(length = 0), minlen = 600,
  norm = FALSE)

mcufd_enrich(cFres, n = NA_real_, rank = "Phylum", plot = FALSE,
  pthresh = NA_real_, fname = NA_character_, units = "in",
  width = 10, height = 7, dpi = 600)

\S4method{mcufd}{codonFreq}(cFobj, exclude = character(length = 0),
  minlen = 600, norm = FALSE)
}
\arguments{
\item{cFobj}{An object of class \code{codonFreq}.}

\item{exclude}{A character vector of codons to be excluded from comparisons.}

\item{minlen}{Numeric, the minimum length of sequence (in codons) to be
included in the analysis. Default = 500.}

\item{norm}{Logical, should the data be normalised? If TRUE, a normalisation
step will be performed. Default = FALSE.}

\item{cFres}{List of data frames containing MCUFD results.}

\item{n}{Numeric, the number of top-ranked reference taxa to be plotted
per input sequence. Default = 100.}

\item{rank}{Character, taxonomic rank to be used for categorisation of the
CUFD hit sequences. Options are "Domain", "Kingdom", and "Phylum".
Default = "Phylum".}

\item{plot}{Logical, should the enrichment results be plotted?
Default = FALSE.}

\item{pthresh}{Numeric, adjusted p-value threshold to be used for subsetting
the data if plot == TRUE.}

\item{fname}{Character, name of the output file (used if plot == TRUE).}

\item{units}{Numeric, units to be used for defining the plot size.
Options are "in" (default), "cm", and "mm".}

\item{width}{Numeric, width of the figure (in \code{units}).}

\item{height}{Numeric, height of the figure (in \code{units}).}

\item{dpi}{Numeric, resolution of the figure (default = 600).}
}
\value{
A list of data frames containing enrichment results.

A list of data framse containing values which indicate the degree of
   codon usage similarity between the sequences used to construct the
   \code{codonFreq} object and those used to construct the reference
   database. The reference database entries are in rows and the
   \code{codonFreq} sequences are in columns.
}
\description{
Compare taxon proportions in the top n hits with those in the full set of
   taxa.

Compare codon usage between sequences in a \code{codonFreq} object with those
   in a reference database. Mean codon usage frequency difference
   (\code{MCUFD}) is calculated as described in
   \href{https://goo.gl/7sGZeE}{Stedman et al. (2013)}.
}
\examples{
## To be added

}
