% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCUFD.R
\name{MCUFD_enrich,list-method}
\alias{MCUFD_enrich,list-method}
\title{MCUFD enrichment testing}
\usage{
\S4method{MCUFD_enrich}{list}(
  cFres,
  n = NA_real_,
  rank = "Phylum",
  save = FALSE,
  pthresh = NA_real_,
  fname = NA_character_,
  units = "in",
  width = 10,
  height = 7,
  dpi = 600,
  ptype = "heatmap",
  outtab = NULL
)
}
\arguments{
\item{cFres}{List of data frames containing MCUFD results.}

\item{n}{Numeric, the number of top-ranked reference taxa to be plotted
per input sequence. Default = 100.}

\item{rank}{Character, taxonomic rank to be used for categorisation of the
CUFD hit sequences. Options are "Domain", "Kingdom", and "Phylum".
Default = "Phylum".}

\item{save}{Logical, should the enrichment plot be saved to file?
Default = FALSE.}

\item{pthresh}{Numeric, adjusted p-value threshold to be used for subsetting
the data if plot == TRUE. Default = NA.}

\item{fname}{Character, name of the output file (used if plot == TRUE).}

\item{units}{Numeric, units to be used for defining the plot size.
Options are "in" (default), "cm", and "mm".}

\item{width}{Numeric, width of the figure (in \code{units}).}

\item{height}{Numeric, height of the figure (in \code{units}).}

\item{dpi}{Numeric, resolution of the figure (default = 600).}

\item{ptype}{Character, type of plot to make (if plot == TRUE).  Options are
"heatmap" (default) and "dotplot".}

\item{outtab}{Character (optional); name of file to which enrichment test
results will be saved. Format is tab-delimited.}
}
\value{
A \code{ggplot} object.
}
\description{
Compare taxon proportions in the top n hits with those in the full set of
   taxa.
}
\examples{
   virusSet <- readSeq(example = TRUE)
   virusCF <- codonFreq(virusSet)
   exclCod <- c("ATT", "TGT")
   MCUFD_tmp <- MCUFD(virusCF, exclude = exclCod, norm = TRUE)
   enrich_tmp <- MCUFD_enrich(
       MCUFD_tmp, n = 100, rank = "Phylum", save = FALSE, pthresh = 0.01
   )

}
