% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-codonFreq.R
\docType{methods}
\name{normalise}
\alias{normalise}
\title{Normalise the codon frequencies in a codonFreq object by amino acid}
\usage{
\S4method{normalise}{codonFreq}(object)
}
\arguments{
\item{object}{An object of class \code{codonFreq}.}
}
\value{
An object of class \code{codonFreq}, containing normalised
   data in which the sum of the relative frequencies per amino acid for a
   given sequence is 1. The Standard Genetic code is currently the only
   code used for normalisation.
}
\description{
Normalise the codon frequencies in a codonFreq object by amino acid
}
