% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codonFreq-class.R
\docType{methods}
\name{nseq}
\alias{nseq}
\alias{nseq,codonFreq-method}
\title{Number of sequences in a \code{codonFreq} object.}
\usage{
nseq(object)

\S4method{nseq}{codonFreq}(object)
}
\arguments{
\item{object}{An object of class \code{codonFreq}.}
}
\value{
Numeric, the number of sequences in the \code{codonFreq} object.
}
\description{
Returns the number of sequences in a \code{codonFreq} object.
}
\section{Methods (by class)}{
\itemize{
\item \code{codonFreq}: 
}}

