% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/PCA.R
\docType{methods}
\name{PCA}
\alias{PCA}
\alias{predict_PCA}
\title{PCA of codon usage by taxon}
\usage{
PCA(exclude = character(length = 0), minlen = 600, norm = FALSE,
  rank = "Phylum", corCut = 0.9, includeTax = NULL)

predict_PCA(cFobj, pcaObj, rank = "Phylum", minlen = 600,
  fname = NA_character_, units = "in", width = 10, height = 7,
  dpi = 600, norm = FALSE, plot = FALSE, identifier = NULL,
  includeTax = NULL)

\S4method{PCA}{ANY}(exclude = character(length = 0), minlen = 600,
  norm = FALSE, rank = "Phylum", corCut = 0.9, includeTax = NULL)
}
\arguments{
\item{exclude}{A character vector of codons to be excluded from comparisons.}

\item{minlen}{Numeric, the minimum length of sequence (in codons) to be
included in the analysis. Default = 500.}

\item{norm}{Logical, should the codon abundances be normalised? If TRUE,
codon abundances will be converted to codon bias scores, such that the sum
of scores for each amino acid sum to 1. Default = FALSE.}

\item{rank}{Character, taxonomic rank to be used for categorisation. Options
are "Domain", "Kingdom", and "Phylum". Default = "Phylum".}

\item{corCut}{Numeric, correlation cutoff used for dropping codons that
exhibit collinearity. Default = 0.9.}

\item{includeTax}{Character, optional vector of taxa (at the specified rank)
to be included in the PCA. If not supplied, all taxa will be used.}
}
\value{
An object of class \code{prcomp}.
}
\description{
Perform PCA on codon usage in a reference
   database and sequences of unknown taxonomic affinity.
}
\examples{
    PCA_tmp <- PCA(exclude = exclCod, rank = "Phylum", minlen = 600)
    class(PCA_tmp)

}
