% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-codonFreq.R
\name{codonPlot,codonFreq-method}
\alias{codonPlot,codonFreq-method}
\alias{biasPlot,codonFreq-method}
\alias{gcPlot,codonFreq-method}
\title{Plot the relative frequencies of codons in a \code{codonFreq} object.}
\usage{
\S4method{codonPlot}{codonFreq}(
  object,
  fname = NA_character_,
  units = "in",
  width = 10,
  height = 7,
  dpi = 600,
  groups = NULL,
  ptype = "boxplot",
  order = "median",
  colour = 1,
  suppress_x_txt = FALSE,
  suppress_y_title = FALSE,
  label = NULL,
  highlight = NULL,
  save = FALSE
)

\S4method{biasPlot}{codonFreq}(
  object,
  fname = NA_character_,
  units = "in",
  width = 10,
  height = 7,
  dpi = 600,
  groups = NULL,
  aa = NULL,
  norm = FALSE,
  label = NULL,
  colours = NULL,
  suppress_y_txt = FALSE,
  suppress_y_title = FALSE,
  legend = TRUE,
  ylim = NULL,
  save = FALSE
)

\S4method{gcPlot}{codonFreq}(
  object,
  fname = NA_character_,
  units = "in",
  width = 10,
  height = 7,
  dpi = 600,
  groups = NULL,
  aa = NULL,
  norm = FALSE,
  label = NULL,
  colours = NULL,
  suppress_y_txt = FALSE,
  suppress_y_title = FALSE,
  legend = TRUE,
  xlim = NULL,
  outtab = NULL,
  save = FALSE
)
}
\arguments{
\item{object}{A \code{codonFreq} object.}

\item{fname}{Character, name of figure generated.}

\item{units}{Numeric, units to be used for defining the plot size.
Options are "in" (default), "cm", and "mm".}

\item{width}{Numeric, width of the figure (in \code{units}).}

\item{height}{Numeric, height of the figure (in \code{units}).}

\item{dpi}{Numeric, resolution of the figure (default = 600).}

\item{groups}{Character, optional vector giving groups for each of the
sequences in the \code{codonFreq} object. If not supplied, the sequences
will be treated as a single group.}

\item{ptype}{Character, type of plot to make.  Options are "boxplot"
(default) and "jitter".}

\item{order}{Character, determines how codons are order. The default
(median) sorts codons by median usage in the full dataset. An alternative
option is "alphabetical". A vector of (64) codons can also be passed,
which will be used for ordering, exactly as in the vector.}

\item{colour}{Integer, between 1 and 9, which specifies the colour to be used
from the `Set1` palette in the `RColorBrewer` package. Only applies to
non-grouped plots. Default = 1 (red).}

\item{suppress_x_txt}{Logical, suppress x axis labels? Default = FALSE.}

\item{suppress_y_title}{Logical, suppress y axis title? Default = FALSE.}

\item{label}{Character, optional label to add to the plotting area.}

\item{highlight}{Character (optional) codons in this vector will be
highlighted.}

\item{save}{Logical, save the plot to file? Default = FALSE.}

\item{aa}{Character, amino acid for which codon bias plot will be made
(must be specified using single letter IUPAC code).}

\item{norm}{Logical, should the data be normalised? If TRUE, a normalisation
step will be performed. Default = FALSE.}

\item{colours}{Optional vector of same length as `groups`, containing
integers between 1 and 9, which specifies the colours to be used from the
`Set1` palette in the RColorBrewer package. By default, the Set1 order is
used.}

\item{suppress_y_txt}{Logical, suppress y axis labels? Default = FALSE.}

\item{legend}{Logical, plot a legend? Default = TRUE.}

\item{ylim}{Numeric vector, gives the y-axis limits (if not supplied, they
will be chosen based on the data).}

\item{xlim}{Numeric vector, gives the x-axis limits (if not supplied, they
will be chosen based on the data).}

\item{outtab}{Character (optional); name of file to which GC3 data will be
saved. Format is tab-delimited.}
}
\value{
A \code{ggplot} object.

A \code{ggplot} object.

A \code{ggplot} object.
}
\description{
Plot the relative frequencies of codons in a \code{codonFreq} object.

Boxplots of normalised codon usage by amino acid (i.e codon bias) in a
   \code{codonFreq} object.

GC3 plots: plot the 3rd-codon GC content for sequences in a \code{codonFreq}
  object.
}
