% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/MCUFD.R
\docType{methods}
\name{mcufd_plot}
\alias{mcufd_plot}
\alias{mcufd_plot,list-method}
\title{MCUFD plot}
\usage{
mcufd_plot(cFres, type = "bar", n = NA_real_, rank = "Phylum",
  fname = NA_character_, units = "in", width = 10, height = 7,
  dpi = 600)

\S4method{mcufd_plot}{list}(cFres, type = "bar", n = NA_real_,
  rank = "Phylum", fname = NA_character_, units = "in", width = 10,
  height = 7, dpi = 600)
}
\arguments{
\item{cFres}{List of data frames containing MCUFD results.}

\item{type}{Character, the type of plot to make. Options are "bar" (default)
and "line".}

\item{n}{Numeric, the number of top-ranked reference taxa to be plotted
per input sequence. Default = 100.}

\item{rank}{Character, taxonomic rank to be used for categorisation of the
CUFD hit sequences. Options are "Domain", "Kingdom", and "Phylum".
Default = "Phylum".}

\item{fname}{Character, name of figure generated.}

\item{units}{Numeric, units to be used for defining the plot size.
Options are "in" (default), "cm", and "mm".}

\item{width}{Numeric, width of the figure (in \code{units}).}

\item{height}{Numeric, height of the figure (in \code{units}).}

\item{dpi}{Numeric, resolution of the figure (default = 600).}
}
\value{
A \code{ggplot} object.
}
\description{
Plot the distribution of MCUFD values per taxon
}
