% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/PCA.R
\name{predict_PCA}
\alias{predict_PCA}
\title{PCA plots}
\usage{
predict_PCA(
  cFobj,
  pcaObj,
  rank = "Phylum",
  minlen = 600,
  fname = NA_character_,
  units = "in",
  width = 10,
  height = 7,
  dpi = 600,
  norm = FALSE,
  save = FALSE,
  identifier = NULL,
  includeTax = NULL,
  colours = NULL
)

\S4method{predict_PCA}{codonFreq}(
  cFobj,
  pcaObj,
  rank,
  minlen,
  fname,
  units,
  width,
  height,
  dpi,
  norm,
  save,
  identifier,
  colours
)
}
\arguments{
\item{cFobj}{An object of class \code{codonFreq}.}

\item{pcaObj}{Object of class \code{prcomp}.}

\item{rank}{Character, taxonomic rank to be used for categorisation. Options
are "Domain", "Kingdom", and "Phylum". Default = "Phylum".}

\item{minlen}{Numeric, the minimum length of sequence (in codons) to be
included in the analysis. Default = 500.}

\item{fname}{Character, name of figure generated.}

\item{units}{Numeric, units to be used for defining the plot size.
Options are "in" (default), "cm", and "mm".}

\item{width}{Numeric, width of the figure (in \code{units}).}

\item{height}{Numeric, height of the figure (in \code{units}).}

\item{dpi}{Numeric, resolution of the figure (default = 600).}

\item{norm}{Logical, should the codon abundances be normalised? If TRUE,
codon abundances will be converted to codon bias scores, such that the sum
of scores for each amino acid sum to 1. Default = FALSE.}

\item{save}{Logical, should the enrichment results plot be saved to a file?
Default = FALSE.}

\item{identifier}{Character, optional group label to be assigned to sequences
in the \code{codonFreq} object. If not supplied, each sequence will be
labelled individually on the plot. Maximum number of unique identifiers
= 6.}

\item{includeTax}{Character, optional vector of taxa (at the specified rank)
to be included in the PCA. If not supplied, all taxa will be used.}

\item{colours}{Integer vector, containing values between 1 and 9, which
specifies the colours to be used from the `Set1` palette in the
`RColorBrewer` package.}
}
\value{
A \code{ggplot} object.
}
\description{
Predict taxonomic classifications for sequences in a \code{codonFreq})
object, using linear discriminants from an \code{PCA}) model. Plot
disciminants and predictions.
}
\examples{
    virusSet <- readSeq(example = TRUE)
    virusCF <- codonFreq(virusSet)
    exclCod <- c("ATT", "TGT")
    PCA_tmp <- PCA(exclude = exclCod, rank = "Phylum", minlen = 600)
    predPCA <- predict_PCA(
        virusCF, PCA_tmp, rank = "Phylum", save = TRUE,
        minlen = 600, fname = "PCA_tmp2", height = 5, width = 7
    )

}
