% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codonFreq-class.R
\docType{methods}
\name{seqID}
\alias{seqID}
\alias{seqID,codonFreq-method}
\title{Get the sequence identifiers in a \code{codonFreq} object.}
\usage{
seqID(object)

\S4method{seqID}{codonFreq}(object)
}
\arguments{
\item{object}{An object of class \code{codonFreq}.}
}
\value{
Character, the identifiers of sequences in the \code{codonFreq}
   object.
}
\description{
Returns the sequence identifiers in a \code{codonFreq} object.
}
\section{Methods (by class)}{
\itemize{
\item \code{codonFreq}: 
}}

