% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codonFreq-class.R
\docType{methods}
\name{seqlen}
\alias{seqlen}
\alias{seqlen,codonFreq-method}
\title{Get the lengths of sequences (nucleotides) in a \code{codonFreq} object.}
\usage{
seqlen(object)

\S4method{seqlen}{codonFreq}(object)
}
\arguments{
\item{object}{An object of class \code{codonFreq}.}
}
\value{
Numeric, lengths of sequences (in nucleotides).
}
\description{
Returns the lengths of sequences (nucleotides) in a \code{codonFreq} object.
}
\section{Methods (by class)}{
\itemize{
\item \code{codonFreq}: 
}}

