% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-codonFreq.R
\name{show,codonFreq-method}
\alias{show,codonFreq-method}
\title{Display the data in a \code{codonFreq} object.}
\usage{
\S4method{show}{codonFreq}(object)
}
\arguments{
\item{object}{A \code{codonFreq} object.}
}
\description{
Display the data in a \code{codonFreq} object.
}
