% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/AllGenerics.R,
%   R/codonFreq-class.R, R/codonComparisons.R, R/methods-codonFreq.R
\docType{class}
\name{codonFreq-class}
\alias{codonFreq-class}
\alias{codonFreq}
\alias{getFreqs}
\alias{nseq}
\alias{seqID}
\alias{seqlen}
\alias{codonFreq,DNAStringSet-method}
\alias{show,codonFreq-method}
\alias{getFreqs,codonFreq-method}
\alias{nseq,codonFreq-method}
\alias{seqID,codonFreq-method}
\alias{seqlen,codonFreq-method}
\title{A class for storing relative codon frequencies in user-supplied sequences.}
\usage{
codonFreq(object)

getFreqs(object)

nseq(object)

seqID(object)

seqlen(object)

codonFreq(object)

\S4method{codonFreq}{DNAStringSet}(object)

\S4method{show}{codonFreq}(object)

codonFreq(object)

\S4method{codonFreq}{DNAStringSet}(object)

\S4method{show}{codonFreq}(object)

getFreqs(object)

\S4method{getFreqs}{codonFreq}(object)

nseq(object)

\S4method{nseq}{codonFreq}(object)

seqID(object)

\S4method{seqID}{codonFreq}(object)

seqlen(object)

\S4method{seqlen}{codonFreq}(object)

\S4method{getFreqs}{codonFreq}(object)

\S4method{nseq}{codonFreq}(object)

\S4method{seqID}{codonFreq}(object)

\S4method{seqlen}{codonFreq}(object)
}
\arguments{
\item{object}{An object of class \code{DNAStringSet}.}

\item{object}{A \code{codonFreq} object.}

\item{object}{An object of class \code{DNAStringSet}.}

\item{object}{A \code{codonFreq} object.}

\item{object}{An object of class \code{codonFreq}.}

\item{object}{An object of class \code{codonFreq}.}

\item{object}{An object of class \code{codonFreq}.}

\item{object}{An object of class \code{codonFreq}.}

\item{object}{An object of class \code{codonFreq}.}
}
\value{
A \code{codonFreq} object.

A \code{codonFreq} object.

Matrix, frequencies of codons in the \code{codonFreq} object.
   Codons are in columns and input sequences are in rows.

Numeric, the number of sequences in the \code{codonFreq} object.

Character, the identifiers of sequences in the \code{codonFreq}
   object.

Numeric, lengths of sequences (in nucleotides).

Matrix, frequencies of codons in the \code{codonFreq} object.
   Codons are in columns and input sequences are in rows.

Numeric, the number of sequences in the \code{codonFreq} object.

Character, the identifiers of sequences in the \code{codonFreq}
   object.

Numeric, lengths of sequences (in nucleotides).
}
\description{
Returns the relative frequencies of codons per sequence in a \code{codonFreq}
   object.

Returns the number of sequences in a \code{codonFreq} object.

Returns the sequence identifiers in a \code{codonFreq} object.

Returns the lengths of sequences (nucleotides) in a \code{codonFreq} object.
}
\section{Methods (by generic)}{
\itemize{
\item \code{codonFreq}: 

\item \code{show}: 

\item \code{codonFreq}: 

\item \code{show}: 

\item \code{getFreqs}: 

\item \code{nseq}: 

\item \code{seqID}: 

\item \code{seqlen}: 

\item \code{getFreqs}: Returns the relative frequencies of codons per sequence
in a \code{codonFreq} object.

\item \code{nseq}: Returns the number of sequences in a \code{codonFreq}
object.

\item \code{seqID}: Returns the sequence identifiers in a \code{codonFreq}
object.

\item \code{seqlen}: Returns the lengths of sequences (nucleotides) in a
\code{codonFreq} object.
}}

\section{Slots}{

\describe{
\item{\code{seqID}}{Character vector of sequence identifiers, derived from Fasta
identifier lines.}

\item{\code{codon_perc}}{Matrix containing the relative frequencies of codons; each
sequence is a row and each codon is a column.}

\item{\code{ntlen}}{Numeric vector of sequence lengths (in nucleotides).}

\item{\code{seqID}}{Character vector of sequence identifiers, derived from Fasta
identifier lines.}

\item{\code{codon_perc}}{Matrix containing the relative frequencies of codons; each
sequence is a row and each codon is a column.}

\item{\code{ntlen}}{Numeric vector of sequence lengths (in nucleotides).}

\item{\code{seqID}}{Character vector of sequence identifiers, derived from Fasta
identifier lines.}

\item{\code{codon_perc}}{Matrix containing the relative frequencies of codons; each
sequence is a row and each codon is a column.}

\item{\code{ntlen}}{Numeric vector of sequence lengths (in nucleotides).}
}}

