% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/AllGenerics.R,
%   R/methods-codonFreq.R
\docType{class}
\name{codonFreq-class}
\alias{codonFreq-class}
\alias{codonFreq}
\alias{getFreqs}
\alias{nseq}
\alias{seqID}
\alias{seqlen}
\alias{codonPlot}
\alias{normalise}
\alias{getFreqs,codonFreq-method}
\alias{nseq,codonFreq-method}
\alias{seqID,codonFreq-method}
\alias{seqlen,codonFreq-method}
\title{S4 class for storing relative codon frequencies in user-supplied sequences.}
\usage{
codonFreq(object)

getFreqs(object)

nseq(object)

seqID(object)

seqlen(object)

codonPlot(
  object,
  fname = NA_character_,
  units = "in",
  width = 10,
  height = 7,
  dpi = 600,
  groups = NULL,
  ptype = "boxplot",
  order = "median",
  colour = 1,
  suppress_x_txt = FALSE,
  suppress_y_title = FALSE,
  label = NULL,
  highlight = NULL,
  save = FALSE
)

normalise(object)

\S4method{getFreqs}{codonFreq}(object)

\S4method{nseq}{codonFreq}(object)

\S4method{seqID}{codonFreq}(object)

\S4method{seqlen}{codonFreq}(object)
}
\arguments{
\item{object}{An object of class \code{codonFreq}.}
}
\value{
Matrix, frequencies of codons in the \code{codonFreq} object.
   Codons are in columns and input sequences are in rows.

Numeric, the number of sequences in the \code{codonFreq} object.

Character, the identifiers of sequences in the \code{codonFreq}
   object.

Numeric, lengths of sequences (in codons).
}
\description{
S4 class for storing relative codon frequencies in user-supplied sequences.
}
\section{Methods (by generic)}{
\itemize{
\item \code{getFreqs}: Returns the relative frequencies of codons per sequence
in a \code{codonFreq} object.

\item \code{nseq}: Returns the number of sequences in a \code{codonFreq}
object.

\item \code{seqID}: Returns the sequence identifiers in a \code{codonFreq}
object.

\item \code{seqlen}: Returns the lengths of sequences (in codons) in a
\code{codonFreq} object.
}}

\section{Slots}{

\describe{
\item{\code{seqID}}{Character vector of sequence identifiers, derived from Fasta
identifier lines.}

\item{\code{freq}}{Matrix containing the relative frequencies of codons; each
sequence is a row and each codon is a column.}

\item{\code{ncod}}{Numeric vector of sequence lengths (in codons).}
}}

