% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/MCUFD.R
\docType{methods}
\name{MCUFD}
\alias{MCUFD}
\alias{MCUFD_enrich}
\alias{MCUFD_plot}
\title{Calculate mean codon usage similarity.}
\usage{
MCUFD(cFobj, exclude = character(length = 0), minlen = 600,
  norm = FALSE)

MCUFD_enrich(cFres, n = NA_real_, rank = "Phylum", save = FALSE,
  pthresh = NA_real_, fname = NA_character_, units = "in",
  width = 10, height = 7, dpi = 600, ptype = "heatmap",
  outtab = NULL)

MCUFD_plot(cFres, type = "bar", n = NA_real_, rank = "Phylum",
  fname = NA_character_, units = "in", width = 10, height = 7,
  dpi = 600)

\S4method{MCUFD}{codonFreq}(cFobj, exclude = character(length = 0),
  minlen = 600, norm = FALSE)
}
\arguments{
\item{cFobj}{An object of class \code{codonFreq}.}

\item{exclude}{A character vector of codons to be excluded from comparisons.}

\item{minlen}{Numeric, the minimum length of sequence (in codons) to be
included in the analysis. Default = 500.}

\item{norm}{Logical, should the data be normalised? If TRUE, a normalisation
step will be performed. Default = FALSE.}
}
\value{
A list of data.frames containing values which indicate the degree of
   codon usage similarity between the sequences used to construct the
   \code{codonFreq} object and those used to construct the reference
   database. The reference database entries are in rows and the
   \code{codonFreq} sequences are in columns.
}
\description{
Compare codon usage between sequences in a \code{codonFreq} object with those
   in a reference database. Mean codon usage frequency difference
   (\code{MCUFD}) is calculated as described in
   \href{https://goo.gl/7sGZeE}{Stedman et al. (2013)}.
}
\examples{
   MCUFD_tmp2 <- MCUFD(tmp2norm, exclude = exclCod)
   range(MCUFD_tmp2[[1]]$MCUFD)
   table(MCUFD_tmp2[[1]]$Kingdom, useNA = "always")
   MCUFD_tmp2[[1]]$Species[1:10]

}
