% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA.R, R/MCUFD.R
\docType{methods}
\name{predict_LDA,codonFreq-method}
\alias{predict_LDA,codonFreq-method}
\alias{MCUFD_plot,list-method}
\title{LDA plots}
\usage{
\S4method{predict_LDA}{codonFreq}(cFobj, ldaObj, rank = "Phylum",
  minlen = 600, fname = NA_character_, units = "in", width = 10,
  height = 7, dpi = 600, norm = FALSE, plot = FALSE,
  identifier = NA_character_)

\S4method{MCUFD_plot}{list}(cFres, type = "bar", n = NA_real_,
  rank = "Phylum", fname = NA_character_, units = "in", width = 10,
  height = 7, dpi = 600, save = FALSE)
}
\arguments{
\item{cFobj}{An object of class \code{codonFreq}.}

\item{ldaObj}{Object of class \code{lda} - produced using the LDA() function.}

\item{rank}{Character, taxonomic rank to be used for categorisation of the
CUFD hit sequences. Options are "Domain", "Kingdom", and "Phylum".
Default = "Phylum".}

\item{minlen}{Numeric, the minimum length of sequence (in codons) to be
included in the analysis. Default = 500.}

\item{fname}{Character, name of figure generated.}

\item{units}{Numeric, units to be used for defining the plot size.
Options are "in" (default), "cm", and "mm".}

\item{width}{Numeric, width of the figure (in \code{units}).}

\item{height}{Numeric, height of the figure (in \code{units}).}

\item{dpi}{Numeric, resolution of the figure (default = 600).}

\item{norm}{Logical, should the codon abundances be normalised? If TRUE,
codon abundances will be converted to codon bias scores, such that the sum
of scores for each amino acid sum to 1. Default = FALSE.}

\item{plot}{Logical, should the enrichment results be plotted?
Default = FALSE.}

\item{identifier}{Character, optional group label to be assigned to sequences
in the \code{codonFreq} object. If not supplied, each sequence will be
labelled individually on the plot.}

\item{cFres}{List of data frames containing MCUFD results.}

\item{type}{Character, the type of plot to make. Options are "bar" (default)
and "line".}

\item{n}{Numeric, the number of top-ranked reference taxa to be plotted
per input sequence. Default = 100.}

\item{save}{Logical, should the figure be saved to file? Default = FALSE.}

\item{rank}{Character, taxonomic rank to be used for categorisation of the
CUFD hit sequences. Options are "Domain", "Kingdom", and "Phylum".
Default = "Phylum".}

\item{fname}{Character, name of figure generated.}

\item{units}{Numeric, units to be used for defining the plot size.
Options are "in" (default), "cm", and "mm".}

\item{width}{Numeric, width of the figure (in \code{units}).}

\item{height}{Numeric, height of the figure (in \code{units}).}

\item{dpi}{Numeric, resolution of the figure (default = 600).}
}
\value{
A \code{ggplot} object.

A \code{ggplot} object.
}
\description{
Predict taxonomic classifications for sequences in a \code{codonFreq})
object, using linear discriminants from an \code{lda}) model. Plot
disciminants and predictions.

Plot the distribution of MCUFD values per taxon
}
\examples{
   predLDA <- predict_LDA(
       tmp2norm, LDA_tmp2, rank = "Phylum", plot = TRUE,
       minlen = 600, fname = "lda_tmp2", height = 5, width = 7
   )
   head(sort(table(predLDA$class), decreasing = TRUE))

   MCUFD_plot(
       cFres, type = "bar", save = TRUE, fname = "MCUFD_bar_kingdom", n = 100,
       rank = "Kingdom"
   )

}
