% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSeq.R
\name{readSeq}
\alias{readSeq}
\title{Read nucleotide sequences from a file.}
\usage{
readSeq(file = character(), ...)
}
\arguments{
\item{file}{Character, path to a file containing one or more protein-coding
DNA sequences in fasta format.}

\item{...}{Other arguments passed to readDNAStringSet or readRNAStringSet.}
}
\value{
Returns a \code{DNAStringSet} object.
}
\description{
Reads a fasta-format \code{file} containing one or more protein-coding
    DNA sequences.
}
\examples{
readSeq(file = "../inst/extdata/example.fna")

}
